fix: some fixes
