+++ webapp/src/components/AssetImage/AssetImage.container.ts
@@ -3,6 +3,7 @@ import { Avatar } from '@dcl/schemas'
 import { connect } from 'react-redux'
 import { RootState } from '../../modules/reducer'
 import { getWallet } from '../../modules/wallet/selectors'
+import { getItem } from '../../modules/item/utils'
 import {
   getIsTryingOn,
   getIsPlayingEmote,
@@ -12,18 +13,26 @@ import {
   setIsTryingOn,
   setWearablePreviewController
 } from '../../modules/ui/preview/actions'
-import { getItem } from '../../modules/item/selectors'
+import { getData as getItems } from '../../modules/item/selectors'
 import {
   MapStateProps,
   MapDispatchProps,
-  MapDispatch
+  MapDispatch,
+  OwnProps
 } from './AssetImage.types'
 import AssetImage from './AssetImage'
 
-const mapState = (state: RootState): MapStateProps => {
+const mapState = (state: RootState, ownProps: OwnProps): MapStateProps => {
   const profiles = getProfiles(state)
   const wallet = getWallet(state)
   let avatar: Avatar | undefined = undefined
+  const items = getItems(state)
+  const item = getItem(
+    ownProps.asset.contractAddress,
+    ownProps.asset.itemId,
+    items
+  )
+
   if (wallet && !!profiles[wallet.address]) {
     const profile = profiles[wallet.address]
     avatar = profile.avatars[0]
@@ -33,7 +42,7 @@ const mapState = (state: RootState): MapStateProps => {
     wearableController: getWearablePreviewController(state),
     isTryingOn: getIsTryingOn(state),
     isPlayingEmote: getIsPlayingEmote(state),
-    item: getItem(state)
+    item
   }
 }
