+++ webapp/src/modules/item/selectors.ts
@@ -13,7 +13,6 @@ export const getState = (state: RootState) => state.item
 export const getData = (state: RootState) => getState(state).data
 export const getError = (state: RootState) => getState(state).error
 export const getLoading = (state: RootState) => getState(state).loading
-export const getItem = (state: RootState) => getState(state).item
 
 export const isFetchingItem = (
   state: RootState,
